Minor change in build_win_installer.bat.

+++ build_win_installer.bat
@@ -1,5 +1,5 @@
 @ECHO OFF
-rd /s /q build
+if exist build rd /s /q build
 rd /s /q dist
 python setup.py bdist_wininst
 rd /s /q build
